+++ access-control-app/src/cmr/access_control/data/access_control_index.clj
@@ -1,19 +1,20 @@
 (ns cmr.access-control.data.access-control-index
   "Performs search and indexing of access control data."
   (:require
-    [clojure.edn :as edn]
-    [clojure.string :as str]
-    [cmr.access-control.data.acls :as acls]
-    [cmr.access-control.services.acl-service :as acl-service]
-    [cmr.common-app.services.search.elastic-search-index :as esi]
-    [cmr.common-app.services.search.query-to-elastic :as q2e]
-    [cmr.common.lifecycle :as l]
-    [cmr.common.log :refer [info debug]]
-    [cmr.common.services.errors :as errors]
-    [cmr.common.util :as util]
-    [cmr.elastic-utils.index-util :as m :refer [defmapping defnestedmapping]]
-    [cmr.transmit.metadata-db2 :as mdb]
-    [cmr.umm.acl-matchers :as acl-matchers]))
+   [clojure.edn :as edn]
+   [clojure.string :as str]
+   [cmr.access-control.data.acls :as acls]
+   [cmr.access-control.services.acl-service :as acl-service]
+   [cmr.common-app.services.search.elastic-search-index :as esi]
+   [cmr.common-app.services.search.query-to-elastic :as q2e]
+   [cmr.common.lifecycle :as l]
+   [cmr.common.log :refer [info debug]]
+   [cmr.common.services.errors :as errors]
+   [cmr.common.util :as util :refer [defn-timed]]
+   [cmr.elastic-utils.index-util :as m :refer [defmapping defnestedmapping]]
+   [cmr.transmit.metadata-db :as mdb-legacy]
+   [cmr.transmit.metadata-db2 :as mdb]
+   [cmr.umm.acl-matchers :as acl-matchers]))
 
 (defmulti index-concept
   "Indexes the concept map in elastic search."
@@ -23,6 +24,15 @@
 (defmethod index-concept :default
   [context concept-map])
   ;; Do nothing
+
+(defn-timed reindex-groups
+  "Fetches and indexes all groups"
+  [context]
+  (info "Reindexing all groups")
+  (doseq [group-batch (mdb-legacy/find-in-batches context :access-group 100 {})
+          group group-batch]
+    (index-concept context group))
+  (info "Reindexing all groups complete"))
 
 (defmulti delete-concept
   "Deletes the concept map in elastic search."