+++ access-control-app/src/cmr/access_control/data/acl/validation.clj
@@ -1,1 +1,0 @@
-(ns cmr.access-control.data.acl.validation)