+++ system-int-test/test/cmr/system_int_test/ingest/collection_validation_test.clj
@@ -97,7 +97,7 @@
    (let [collection (assoc (dc/collection coll-attributes) :native-id (:native-id coll-attributes))
          provider-id (get coll-attributes :provider-id "PROV1")
          response (d/ingest provider-id collection options)]
-     (is (= {:status 200} (select-keys response [:status :errors]))))))
+     (is (#{{:status 200} {:status 201}} (select-keys response [:status :errors]))))))
 
 (defn assert-valid-keywords
   [coll-attributes]
@@ -610,7 +610,7 @@
     (are3 [format collection warning-message]
           (do
             (let [response (d/ingest "PROV1" collection {:format format})]
-              (is (= 200 (:status response)))
+              (is (#{200 201} (:status response)))
               (is (= warning-message (:warnings response))))
             (let [response (ingest/validate-concept (dc/collection-concept collection format))]
               (is (= 200 (:status response)))
@@ -632,13 +632,13 @@
           :iso-smap (dc/collection-smap {}) "object has missing required properties ([\"DataCenters\",\"Platforms\",\"ProcessingLevel\",\"RelatedUrls\",\"ScienceKeywords\",\"SpatialExtent\",\"TemporalExtents\"])"
 
           "Multiple Warnings"
-          :dif10 (dc/collection-dif10 {:product (dc/product {:short-name (apply str (repeat 81 "x"))
+          :dif10 (dc/collection-dif10 {:product (dc/product {:short-name (apply str (repeat 121 "x"))
                                                              :processing-level-id "1"})
                                        :platforms [(dc/platform {:short-name (apply str (repeat 81 "x"))})]})
-              "/Platforms/0/ShortName string \"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx\" is too long (length: 81, maximum allowed: 80)/ShortName string \"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx\" is too long (length: 81, maximum allowed: 80)")
+              "/Platforms/0/ShortName string \"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx\" is too long (length: 81, maximum allowed: 80)/ShortName string \"xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx\" is too long (length: 121, maximum allowed: 120)")
    (testing "Warnings returned in JSON format"
      (let [response (d/ingest "PROV1" (dc/collection-dif10 {}) {:format :dif10 :accept-format :json})]
-       (is (= 200 (:status response)))
+       (is (= 201 (:status response)))
        (is (= ["object has missing required properties ([\"ProcessingLevel\"])"] (:warnings response)))))))
 
 (comment