+++ CHANGELOG.md
@@ -109,10 +109,18 @@ Items to include in each issue
 
 ## 1.60
 
+* CMR-3350 - As an Ops user, I should have a way of determining via Splunk if an ingest request is new or updating an existing granule
+  * Client Impacts
+    * CMR changed http status code for ingesting a new concept from 200 to 201, updating or deleting an existing concept still returns http status code 200.
 * CMR-3355 - Changed collection temporal indexing to use umm-spec-lib
   * Client Impacts
     * CMR enabled unique granule ur validation during granule ingest so that ingest of granule with a granule ur that is used by a different non-deleted granule within the same provider will fail.
 * CMR-3340 Translate UMM ISOTopicCategories to ISO 19115-2 (MENDS)
   * Updated UMM JSON schema to 1.7 which changed ISOTopicCategory type from string to ISOTopicCategoryEnum.
 
+## 1.61
 
+* CMR-2717 - Changed legacy services to read groups from CMR access control
+  * Needs config v341
+  * Need to manually synchronize groups before deployment.
+  * The CMR deployment must happen first.